✅ (dashboard): Ratchet quarantine on identities, and load it locally too
Two gaps in the count-based ceiling:

- It counted entries, not identities, so a 1-for-1 swap — retire a fixed spec,
  quarantine a newly-broken one — passed silently and let the list rot at a
  constant size while the number looked reassuring. Now asserts the live list
  is a subset of a frozen baseline: removals stay free, additions and swaps
  fail loudly. Subset subsumes a count, since a subset is never larger.
- "Evaluated on every config load, local and CI" was true only of the CI
  config; `pnpm test:e2e` never loads it. The list and its ratchet move to
  playwright.quarantine.ts, imported by both configs, so the claim now holds.

Zero headroom for additions is deliberate: the next genuinely-quarantined spec
should have to edit the baseline in the same commit, where a reviewer sees it.
